fix: move `mockGraphQLResponse`
