+++ apps/spruce/playwright/helpers/index.ts
@@ -164,4 +164,5 @@ export {
   login,
   logout,
   clickCheckboxByLabel,
+  mockGraphQLResponse,
 } from "@evg-ui/playwright-config/helpers";
