+++ apps/spruce/playwright/tests/task/subscription_modal.spec.ts
@@ -1,7 +1,10 @@
 import { Page } from "@playwright/test";
 import { test, expect } from "../../fixtures";
-import { selectOption, validateToast } from "../../helpers";
-import { mockGraphQLResponse } from "../../utils";
+import {
+  selectOption,
+  validateToast,
+  mockGraphQLResponse,
+} from "../../helpers";
 
 const TASK_ROUTE =
   "/task/evergreen_ubuntu1604_test_model_patch_5e823e1f28baeaa22ae00823d83e03082cd148ab_5e4ff3abe3c3317e352062e4_20_02_21_15_13_48/logs";
