+++ packages/playwright-config/helpers/index.ts
@@ -71,3 +71,41 @@ export const clickCheckboxByLabel = async (page: Page, name: string) => {
     await checkbox.locator("..").locator("label").click();
   }
 };
+
+type ResponseData = {
+  errors: Array<{
+    message: string;
+    path: string[];
+    extensions: { code: string };
+  }> | null;
+  data: unknown;
+};
+
+/**
+ * Helper function to mock GraphQL response.
+ * @param page - The Playwright page object
+ * @param operationName - name of the operation to mock
+ * @param responseData - The mock response data
+ */
+export async function mockGraphQLResponse(
+  page: Page,
+  operationName: string,
+  responseData: ResponseData,
+) {
+  await page.route("**/graphql/query", async (route) => {
+    const request = route.request();
+    const postData = request.postDataJSON();
+    if (postData?.operationName === operationName) {
+      await route.fulfill({
+        status: 200,
+        contentType: "application/json",
+        body: JSON.stringify({
+          errors: responseData.errors,
+          data: responseData.data,
+        }),
+      });
+    } else {
+      await route.continue();
+    }
+  });
+}
